Create two separate builds of the hourglasses.
The artrayd hourglass is nice but I tried it in two forms - one with
a border and one without. I think I prefer the bordered one, but I
don't like that it borders the inside as well. However, it offers a
little insight into how you change the configuration to build different
types of hourglasses.

The 'makehourglass.py' source is still in the root of the repository,
because it is shared.