Updated the other hourglasses to use the shared configuration.
All 3 hourglasses can now be built, and share the same RISC OS
makefiles to do so. This means that there are now 3 hourglasses
that I can run from the same builder. More if I build them, which
is pretty easy to do if you just work out the commands for the
hourglass frame extraction.