Update the bordered hourglass to also be filled.
I've finally worked out the sequence of operations needed in ImageMagick
to fill the inside of the hourglass. This isn't hard per-se, but getting
the right combination of operations and masking settings isn't obvious,
so this is a bit fiddly.

It does require an extra colour to be introduced, so there's an extra
'generatedpalette' variable for the colours that we put into the image
before we convert to the shape.py version.