Fix for problems running under ImageMagick 6.
It seems that ImageMagick 7 isn't actually bundled into Ubuntu yet,
or even available as a PPA. So, for now, the scripts have to be
tested under ImageMagick 6 and 7.

The scripts have been updated to generate an example file for each
hourglass, to make it easier to see what those hourglasses look
like.