Add support for drawing larger hourglasses.
One of the hourglasses has been made larger to see how well this works.
It is not ideal that there is only a single size of hourglass supported
by the Hourglass module, so this ends up being a hardcoded effect.
Maybe a configuration could make this optional in the future.