Add support for building the hourglass modules for RISC OS 64.
We now have support for generating 64bit assembler for the hourglass.
It's arguable (easily) that I should not have written the hourglass
building code in assembler, but should have just used C throughout.
Hey-ho. In the future, maybe we'll change this.

It's somewhat amusing that to make the C-based hourglass work, we
generate ObjAsm style assembler using Python, which is then parsed
by the ObjAsm-to-GAS format Perl tool, and then assembled by GAS.